requirements.txt
PyQt6>=6.4.0
PyOpenGL>=3.1.7
websockets>=12.0
numpy>=1.24.0
